requirements.txt
Cython
tensorflow-gpu
